Remove the dependency on collections package
The collections package we are using seems to be an obsolete pollyfill,
so depending on how you use this package, there is a problem that pollutes
the global scope and breaks the runtime. For example, there seems to be
a bug in the Object.values() function of collections. These modern
ECMAScript features are already available as standard in many browsers
and Node.js, so it is better to use the native standard features rather
than fixing collections package.

Instead of collections/fast-map, it will change it to use native
JavaScript Maps. These work faster in current browsers and Node.js.