Make oniguruma pluggable
We have changed oniguruma https://github.com/atom/node-oniguruma to
peerDependencies package so that we can use WebAssembly-based onigasm
https://github.com/zikaari/onigasm instead of oniguruma. Unless oniguruma is
installed, we will work with onigasm.

We also added an init() function to allow the use of onigasm. You will
need to call this function at the beginning of your application to
initialize WebAssembly.